requirements.txt
requests
urllib3
python-dotenv
mkdocs-material
mkdocs-git-revision-date-localized-plugin